Extract into tree.ts
